+++ packages/component-video/lib/players/BasicPlayer.js
@@ -31,7 +31,7 @@ var BasicPlayer = function (_AbstractPlayer) {
 	_inherits(BasicPlayer, _AbstractPlayer);
 
 	function BasicPlayer() {
-		var mOptions = arguments.length <= 0 || arguments[0] === undefined ? {} : arguments[0];
+		var mOptions = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};
 
 		_classCallCheck(this, BasicPlayer);
 
@@ -59,23 +59,28 @@ var BasicPlayer = function (_AbstractPlayer) {
 		_this._player = document.createElement('video');
 
 		_this._player.setAttribute('playsinline', 'playsinline');
-		_this._player.setAttribute('webkitplaysinline', 'webkitplaysinline');
 
-		var _this$_options = _this._options;
-		var src = _this$_options.src;
-		var poster = _this$_options.poster;
-		var _this$_options$loop = _this$_options.loop;
-		var loop = _this$_options$loop === undefined ? false : _this$_options$loop;
-		var _this$_options$contro = _this$_options.controls;
-		var controls = _this$_options$contro === undefined ? true : _this$_options$contro;
-		var _this$_options$volume = _this$_options.volume;
-		var volume = _this$_options$volume === undefined ? 1 : _this$_options$volume;
-		var _this$_options$preloa = _this$_options.preload;
-		var preload = _this$_options$preloa === undefined ? 'auto' : _this$_options$preloa;
-		var _this$_options$crossO = _this$_options.crossOrigin;
-		var crossOrigin = _this$_options$crossO === undefined ? null : _this$_options$crossO;
+		var _this$_options = _this._options,
+		    src = _this$_options.src,
+		    poster = _this$_options.poster,
+		    _this$_options$loop = _this$_options.loop,
+		    loop = _this$_options$loop === undefined ? false : _this$_options$loop,
+		    _this$_options$contro = _this$_options.controls,
+		    controls = _this$_options$contro === undefined ? true : _this$_options$contro,
+		    _this$_options$volume = _this$_options.volume,
+		    volume = _this$_options$volume === undefined ? 1 : _this$_options$volume,
+		    _this$_options$preloa = _this$_options.preload,
+		    preload = _this$_options$preloa === undefined ? 'auto' : _this$_options$preloa,
+		    _this$_options$crossO = _this$_options.crossOrigin,
+		    crossOrigin = _this$_options$crossO === undefined ? null : _this$_options$crossO,
+		    _this$_options$muted = _this$_options.muted,
+		    muted = _this$_options$muted === undefined ? false : _this$_options$muted,
+		    _this$_options$playsi = _this$_options.playsinline,
+		    playsinline = _this$_options$playsi === undefined ? true : _this$_options$playsi;
 
 
+		_this.muted = muted;
+		_this.playsinline = playsinline;
 		_this.loop = loop;
 		_this.controls = controls;
 		_this.volume = volume;
@@ -169,7 +174,7 @@ var BasicPlayer = function (_AbstractPlayer) {
 	};
 
 	BasicPlayer.prototype._handlePageVisibility = function _handlePageVisibility() {
-		var remove = arguments.length <= 0 || arguments[0] === undefined ? false : arguments[0];
+		var remove = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : false;
 
 		if (typeof document === 'undefined') return;
 		this._hidden = _hiddenPropertyName2.default;
@@ -184,7 +189,7 @@ var BasicPlayer = function (_AbstractPlayer) {
 	};
 
 	BasicPlayer.prototype._handlePageResize = function _handlePageResize() {
-		var remove = arguments.length <= 0 || arguments[0] === undefined ? false : arguments[0];
+		var remove = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : false;
 
 		if (typeof window === 'undefined') return;
 		if (remove) {
@@ -238,6 +243,13 @@ var BasicPlayer = function (_AbstractPlayer) {
 			return this._player.src;
 		}
 	}, {
+		key: 'playsinline',
+		set: function set(value) {
+			if (value) {
+				this._player.setAttribute('playsinline', '');
+			}
+		}
+	}, {
 		key: 'volume',
 		set: function set(value) {
 			this._player.volume = value;
@@ -279,6 +291,15 @@ var BasicPlayer = function (_AbstractPlayer) {
 		},
 		get: function get() {
 			return this._player.currentTime;
+		}
+	}, {
+		key: 'muted',
+		set: function set(muted) {
+			if (muted) {
+				this._player.setAttribute('muted', '');
+			} else {
+				this._player.removeAttribute('muted', '');
+			}
 		}
 	}, {
 		key: 'duration',