+++ packages/component-video/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.3.3 (January 19, 2017)
+
+### Component Video
+
+#### Basic Player
+
+* Muted option available to autoplay videos on iOS10 and Android (for audio you still need to use the hack of providing a audio file)
+* FIX - `playsinline` attribute wasn't set correctly and doesnt require a prefix.
+
 ## 0.3.1 (January 19, 2017)
 
 ### Component Video